k/client/tests: split direct_consumer_fixture into a lib
To test lib is to be used in the upcoming PR.
