+++ src/v/kafka/client/direct_consumer/tests/BUILD
@@ -1,4 +1,4 @@
-load("//bazel:test.bzl", "redpanda_cc_gtest")
+load("//bazel:test.bzl", "redpanda_cc_gtest", "redpanda_test_cc_library")
 
 redpanda_cc_gtest(
     name = "data_queue_test",
@@ -35,6 +35,31 @@ redpanda_cc_gtest(
     ],
 )
 
+redpanda_test_cc_library(
+    name = "direct_consumer_fixture",
+    srcs = [
+        "direct_consumer_fixture.cc",
+    ],
+    hdrs = [
+        "direct_consumer_fixture.h",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//src/v/cluster/tests:cluster_test_fixture",
+        "//src/v/container:chunked_vector",
+        "//src/v/kafka/client:brokers",
+        "//src/v/kafka/client:cluster",
+        "//src/v/kafka/client:types",
+        "//src/v/kafka/client/direct_consumer",
+        "//src/v/kafka/protocol",
+        "//src/v/kafka/server/tests:kafka_test_utils",
+        "//src/v/model/tests:random",
+        "//src/v/test_utils:gtest",
+        "@googletest//:gtest",
+        "@seastar",
+    ],
+)
+
 redpanda_cc_gtest(
     name = "direct_consumer_fixture_test",
     timeout = "moderate",
@@ -42,17 +67,9 @@ redpanda_cc_gtest(
         "direct_consumer_fixture_test.cc",
     ],
     deps = [
-        "//src/v/base",
-        "//src/v/cluster/tests:cluster_test_fixture",
-        "//src/v/container:chunked_vector",
-        "//src/v/kafka/client:brokers",
-        "//src/v/kafka/client:cluster",
-        "//src/v/kafka/client/direct_consumer",
-        "//src/v/kafka/protocol",
-        "//src/v/kafka/server/tests:kafka_test_utils",
+        ":direct_consumer_fixture",
         "//src/v/model",
         "//src/v/redpanda/tests:fixture",
-        "//src/v/test_utils:gtest",
         "@fmt",
         "@googletest//:gtest",
         "@seastar",
